fix(security): encode request values in every listing, lookup and PDF endpoint
The request, result-entry, print, failed, approval and status listings of
VL, CD4, TB, Hepatitis, COVID-19 and Custom Tests, their PDF generators,
mail helpers and reference grids, and the shared admin and lookup
endpoints put request values straight into SQL. _sanitizeInput() only
cleans HTML, so a filter value, a paging offset or a sort direction could
rewrite the query.

The filter block these endpoints copy from each other is now one map per
endpoint handed to ListingFilterClauseBuilder, which writes each clause
with the value encoded for its kind: escape() for = and LIKE, escapeLike()
for contains searches, inIntList() for id lists, (int) for single ids.
Each map keeps the endpoint's own column and comparison, so a legitimate
value matches the rows it matched before. DataTableUtility::paging()
replaces the copied iDisplayStart/iDisplayLength block with int casts,
and sort directions are held to ASC/DESC. Clauses that do not fit the map
(string-built WHERE, OR blocks, base64 values) are encoded in place.

Endpoints that took identifiers from the request now accept only what
their pages send: the generic lookups (get-data-list, get-data-list-for-
generic) against every caller's table and columns, the per-module
sample-duplicate checks against their form tables, get-sample-type
against the known test types, and the shared checkDuplicate against
plain identifiers. Quoted sample-code lists for PDFs are re-read quote by
quote, so a code containing a comma still prints. The instrument config
file name is pinned with basename().

Differential probes compared each old clause with its new one on real
data (VL 1.6M, EID 14K, COVID-19 53K rows): no count changed. Each module
gets an integration test driving its listings with values that listed
every sample, or broke the query, before this change and still finding
the legitimate sample. check-sql-value-safety now covers 273 endpoints
and flags uncast paging values.

Two visible differences: a whitespace-only filter is ignored instead of
matching nothing, and % or _ typed into a contains search match
literally.

+++ app/covid-19/results/save-covid-19-qc-data.php
@@ -28,7 +28,7 @@ try {
         $exist = false;
         if (isset($_POST['qcDataId']) && $_POST['qcDataId'] != "") {
             /* Suppose while edit they can change the testkit means prev data not needed so we can rease it from DB */
-            $exist = $db->rawQueryOne("SELECT qc_id FROM $tableName1 WHERE qc_id = " . base64_decode((string) $_POST['qcDataId']));
+            $exist = $db->rawQueryOne("SELECT qc_id FROM $tableName1 WHERE qc_id = " . (int) base64_decode((string) $_POST['qcDataId']));
             if (isset($exist) && !empty($exist['qc_id'])) {
                 $db->where("qc_id", $exist['qc_id']);
                 $db->delete($tableName1);
